+++ src/GroupCode/FeatureCollector.java
@@ -73,7 +73,7 @@ public class FeatureCollector {
         nodeEdges.clear();
     }
 
-    public void findGameFeatures(byte[] searchSpace, byte[] prevSearchSpace, int xDim, int yDim, int gameStep, int move, int mctsScore, int nodeID) {
+    public void findGameFeatures(byte[] searchSpace, byte[] prevSearchSpace, int xDim, int yDim, int gameStep, int move, float mctsScore, int nodeID) {
         for (BoardFeatures gameFeature : gameFeatures.values()) {
             if (gameFeature instanceof Moves){
                 gameFeature.findFeatures(prevSearchSpace, xDim, yDim, gameStep, move, mctsScore, nodeID);
